Added MoveAmbiguity class + minor cleanup
