+++ Chess.Benchmark/PerftBenchRunner.cs
@@ -97,8 +97,8 @@ No perft cache:
         public static void Main(string[] args)
         {
             //var summary = BenchmarkRunner.Run<FenBenchmark>();
-            //var summary = BenchmarkRunner.Run<PerftBench>();
-            var summary = BenchmarkRunner.Run<SquareOppositeColorBenchmark>();
+            var summary = BenchmarkRunner.Run<PerftBench>();
+            //var summary = BenchmarkRunner.Run<SquareOppositeColorBenchmark>();
         }
     }
 }
