+++ Chess/Types/MoveExtensions.cs
@@ -43,26 +43,6 @@ namespace Rudz.Chess.Types
     {
         public static readonly Move EmptyMove = new Move();
 
-        private static readonly Dictionary<MoveNotations, Func<Move, IPosition, string>> NotationFuncs = new Dictionary<MoveNotations, Func<Move, IPosition, string>>
-        {
-            {MoveNotations.Fan, ToFan},
-            {MoveNotations.San, ToSan},
-            {MoveNotations.Lan, ToLan},
-            {MoveNotations.Ran, ToRan},
-            {MoveNotations.Uci, ToUci}
-        };
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static string ToNotation(this Move move, IPosition pos, MoveNotations notation = MoveNotations.Fan)
-        {
-            if (move.IsNullMove())
-                return "(none)";
-
-            if (!NotationFuncs.TryGetValue(notation, out var func))
-                throw new InvalidMoveException("Invalid move notation detected.");
-
-            return func(move, pos);
-        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static (bool, Move) Locate(this Move move, IPosition pos)
@@ -74,296 +54,6 @@ namespace Rudz.Chess.Types
             return element == null ? (false, EmptyMove) : (true, element);
         }
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static string ToUci(Move move, IPosition pos = null) => move.ToString();
 
-        /// <summary>
-        /// <para>Converts a move to FAN notation.</para>
-        /// </summary>
-        /// <param name="move">The move to convert</param>
-        /// <param name="pos">The position from where the move exist</param>
-        /// <returns>FAN move string</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static string ToFan(Move move, IPosition pos)
-        {
-            var from = move.GetFromSquare();
-            var to = move.GetToSquare();
-
-            var notation = new StringBuilder(12);
-
-            if (move.IsCastlelingMove())
-                notation.Append(CastlelingExtensions.GetCastlelingString(to, from));
-            else
-            {
-                var pc = move.GetMovingPiece();
-                var pt = pc.Type();
-
-                if (pt != PieceTypes.Pawn)
-                {
-                    notation.Append(pc.GetUnicodeChar());
-                    move.Disambiguation(from, pos, notation);
-                }
-
-                if (move.IsEnPassantMove())
-                {
-                    notation.Append("ep");
-                    notation.Append(from.FileChar());
-                }
-                else if (move.IsCaptureMove())
-                {
-                    if (pt == PieceTypes.Pawn)
-                        notation.Append(from.FileChar());
-                    notation.Append('x');
-                }
-
-                notation.Append(to.ToString());
-
-                if (move.IsPromotionMove())
-                {
-                    notation.Append('=');
-                    notation.Append(move.GetPromotedPiece().GetUnicodeChar());
-                }
-            }
-
-            if (pos.State.InCheck)
-                notation.Append(pos.GetCheckChar());
-
-            return notation.ToString();
-        }
-
-        /// <summary>
-        /// <para>Converts a move to SAN notation.</para>
-        /// </summary>
-        /// <param name="move">The move to convert</param>
-        /// <param name="pos">The position from where the move exist</param>
-        /// <returns>SAN move string</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static string ToSan(this Move move, IPosition pos)
-        {
-            var from = move.GetFromSquare();
-            var to = move.GetToSquare();
-
-            var notation = new StringBuilder(12);
-
-            if (move.IsCastlelingMove())
-                notation.Append(CastlelingExtensions.GetCastlelingString(to, from));
-            else
-            {
-                var pt = move.GetMovingPieceType();
-
-                if (pt != PieceTypes.Pawn)
-                {
-                    notation.Append(move.GetMovingPiece().GetPgnChar());
-                    move.Disambiguation(from, pos, notation);
-                }
-
-                if (move.IsEnPassantMove())
-                {
-                    notation.Append("ep");
-                    notation.Append(from.FileChar());
-                }
-                else if (move.IsCaptureMove())
-                {
-                    if (pt == PieceTypes.Pawn)
-                        notation.Append(from.FileChar());
-
-                    notation.Append('x');
-                }
-
-                notation.Append(to.ToString());
-
-                if (move.IsPromotionMove())
-                {
-                    notation.Append('=');
-                    notation.Append(move.GetPromotedPiece().GetPgnChar());
-                }
-            }
-
-            if (pos.State.InCheck)
-                notation.Append(pos.GetCheckChar());
-
-            return notation.ToString();
-        }
-
-        /// <summary>
-        /// <para>Converts a move to LAN notation.</para>
-        /// </summary>
-        /// <param name="move">The move to convert</param>
-        /// <param name="pos">The position from where the move exist</param>
-        /// <returns>LAN move string</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static string ToLan(Move move, IPosition pos)
-        {
-            var from = move.GetFromSquare();
-            var to = move.GetToSquare();
-
-            var notation = new StringBuilder(12);
-
-            if (move.IsCastlelingMove())
-                notation.Append(CastlelingExtensions.GetCastlelingString(to, from));
-            else
-            {
-                var pt = move.GetMovingPieceType();
-
-                if (pt != PieceTypes.Pawn)
-                    notation.Append(pt.GetPieceChar());
-
-                notation.Append(from.ToString());
-
-                if (move.IsEnPassantMove())
-                {
-                    notation.Append("ep");
-                    notation.Append(from.FileChar());
-                }
-                else if (move.IsCaptureMove())
-                {
-                    if (pt == PieceTypes.Pawn)
-                        notation.Append(from.FileChar());
-
-                    notation.Append('x');
-                }
-                else
-                    notation.Append('-');
-
-                notation.Append(to.ToString());
-
-                if (move.IsPromotionMove())
-                {
-                    notation.Append('=');
-                    notation.Append(move.GetPromotedPiece().GetUnicodeChar());
-                }
-            }
-
-            if (pos.State.InCheck)
-                notation.Append(pos.GetCheckChar());
-
-            return notation.ToString();
-        }
-
-        /// <summary>
-        /// <para>Converts a move to RAN notation.</para>
-        /// </summary>
-        /// <param name="move">The move to convert</param>
-        /// <param name="pos">The position from where the move exist</param>
-        /// <returns>RAN move string</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static string ToRan(Move move, IPosition pos)
-        {
-            var from = move.GetFromSquare();
-            var to = move.GetToSquare();
-
-            var notation = new StringBuilder(12);
-
-            if (move.IsCastlelingMove())
-                notation.Append(CastlelingExtensions.GetCastlelingString(to, from));
-            else
-            {
-                var pt = move.GetMovingPieceType();
-
-                if (pt != PieceTypes.Pawn)
-                    notation.Append(pt.GetPieceChar());
-
-                notation.Append(from.ToString());
-
-                if (move.IsEnPassantMove())
-                {
-                    notation.Append("ep");
-                    notation.Append(from.FileChar());
-                }
-                else if (move.IsCaptureMove())
-                {
-                    if (pt == PieceTypes.Pawn)
-                        notation.Append(from.FileChar());
-
-                    notation.Append('x');
-                    notation.Append(move.GetCapturedPiece().Type().GetPieceChar());
-                }
-                else
-                    notation.Append('-');
-
-                notation.Append(to.ToString());
-
-                if (move.IsPromotionMove())
-                {
-                    notation.Append('=');
-                    notation.Append(move.GetPromotedPiece().GetUnicodeChar());
-                }
-            }
-
-            if (pos.State.InCheck)
-                notation.Append(pos.GetCheckChar());
-
-            return notation.ToString();
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static char GetCheckChar(this IPosition pos) => pos.GenerateMoves().Any() ? '+' : '#';
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static MoveAmbiguities Ambiguity(this Move move, BitBoard similarTypeAttacks, IPosition position)
-        {
-            var ambiguity = MoveAmbiguities.None;
-
-            foreach (var square in similarTypeAttacks)
-            {
-                var c = move.GetMovingSide();
-                var pinned = position.GetPinnedPieces(square, c);
-
-                if (similarTypeAttacks & pinned)
-                    continue;
-
-                if (move.GetMovingPieceType() != position.GetPieceType(square))
-                    continue;
-
-                if (position.OccupiedBySide[c.Side] & square)
-                {
-                    if (square.File() == move.GetFromSquare().File())
-                        ambiguity |= MoveAmbiguities.File;
-                    else if (square.Rank() == move.GetFromSquare().Rank())
-                        ambiguity |= MoveAmbiguities.Rank;
-
-                    ambiguity |= MoveAmbiguities.Move;
-                }
-            }
-
-            return ambiguity;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static BitBoard GetSimilarAttacks(this IPosition position, Move move)
-        {
-            var pt = move.GetMovingPieceType();
-
-            return pt == PieceTypes.Pawn || pt == PieceTypes.King
-                ? BitBoards.EmptyBitBoard
-                : move.GetToSquare().GetAttacks(pt, position.Pieces()) ^ move.GetFromSquare();
-        }
-
-        /// <summary>
-        /// Disambiguation.
-        /// <para>If we have more then one piece with destination 'to'.</para>
-        /// <para>Note that for pawns is not needed because starting file is explicit.</para>
-        /// </summary>
-        /// <param name="move">The move to check</param>
-        /// <param name="from">The from square</param>
-        /// <param name="position">The current used position</param>
-        /// <param name="sb">The StringBuilder to append to if needed</param>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static void Disambiguation(this Move move, Square from, IPosition position, StringBuilder sb)
-        {
-            var similarAttacks = position.GetSimilarAttacks(move);
-
-            var ambiguity = move.Ambiguity(similarAttacks, position);
-
-            if (!ambiguity.HasFlagFast(MoveAmbiguities.Move))
-                return;
-
-            if (!ambiguity.HasFlagFast(MoveAmbiguities.File))
-                sb.Append(from.FileChar());
-            else if (!ambiguity.HasFlagFast(MoveAmbiguities.Rank))
-                sb.Append(from.RankChar());
-            else
-                sb.Append(from.ToString());
-        }
     }
 }
